+++ Expanded Cheat Menu/Changelog.md
@@ -7,12 +7,14 @@
 ### Added:
 - Debug>Attach Spots List: Shows list of attaches for use with .ent files.
 - Debug>Object Flags: Shows list of flags set for selected object (also added to Examine>Tools).
+- Debug>Material Properties: Shows list of material settings for use with .mtl files.
 
 ### Changed:
 - Added checkbox for Advanced probes to Add Orbital Probes.
 - The main button area of the selection panel can now be toggled (to make room for the cheats pane).
 - Removed the "mod_texture_" prefix when converting textures to entity files (ConvertImagesToResEntities).
 - Renamed Object Spawner to Entity Spawner.
+- Moved Export CSV from Debug to Game, and added a new export option for graph data.
 
 ## v11.4 (01 Dec 2018)
 ### Changed: